requirements.txt
Pillow==9.5.0
streamlit==1.34.0
numpy==1.26.3
pandas==2.2.2
matplotlib==3.7.1
seaborn==0.12.1
plotly.express==0.4.1
requests==2.27.1
toml==0.10.2
click==8.1.7
wordcloud==1.